+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## natlib 0.1.9 (?)
 
+* ➕ Functions for converting to and from decibels
 * Cut the workaround for `visualViewport` theoretically returning `null`
 
 ## natlib 0.1.8 (2023-01-03)